Add --no-crud option to skip generating CRUD functions
In certain cases it might be not desireable to write CRUD functions by
hand. To avoid conflict this change adds the option to skip generating
those functions. This for example is a workaround for MySQL which
currently is not supported by the generated functions or to support
unsigned integers in id fields.

Related: https://github.com/Wulf/dsync/issues/46
Related: https://github.com/Wulf/dsync/issues/47