Import `diesel::prelude::*` in generated code ignoring unused warning
The diesel::prelude members should always be non-conflicting with diesel
code and hence are safe to import globally. Based on the options used
when generating the may or may not be used by the resulting file and
hence the import is allowed to be unused.

Related: https://github.com/Wulf/dsync/issues/94